+++ src/ZenViewUtil.ts
@@ -1,6 +1,5 @@
 import * as vscode from 'vscode';
 import { ZenViewFile } from './ZenViewFile';
-import * as fs from 'fs';
 
 export class ZenViewUtil {
 
@@ -22,8 +21,22 @@ export class ZenViewUtil {
         return new ZenViewFile(fileUri, collapsibleState, fileType, this.getFileName(fileUri));
     }
 
-    static getFileType(fileUri: vscode.Uri) : vscode.FileType {
-        let isDirectory = fs.lstatSync(fileUri.fsPath).isDirectory();
-        return isDirectory ? vscode.FileType.Directory : vscode.FileType.File;
+    static getFileType(fileUri: vscode.Uri, statSyncFunc: any): vscode.FileType {
+        let fileType: vscode.FileType = vscode.FileType.Unknown;
+        let fileInfo = statSyncFunc(fileUri.fsPath);
+        let isDirectory: boolean = fileInfo.isDirectory();
+        let isSymbolicLink: boolean = fileInfo.isSymbolicLink();
+        let isFile: boolean = fileInfo.isFile();
+
+        if (isSymbolicLink) {
+            fileType = vscode.FileType.SymbolicLink;
+        }
+        else if (isDirectory) {
+            fileType = vscode.FileType.Directory;
+        }
+        else if (isFile) {
+            fileType = vscode.FileType.File;
+        }
+        return fileType;
     }
 }